Add inlined webworker build
