--- packages/doenetml/index.js
@@ -1,17 +1,0 @@
-export { DoenetML } from "./src/DoenetML.jsx";
-
-export { mathjaxConfig } from "./src/Core/utils/math";
-export { cidFromText } from "./src/Core/utils/cid";
-export { retrieveTextFileForCid } from "./src/Core/utils/retrieveTextFile";
-export {
-    calculateOrderAndVariants,
-    determineNumberOfActivityVariants,
-    parseActivityDefinition,
-    returnNumberOfActivityVariantsForCid,
-} from "./src/utils/activityUtils";
-export {
-    serializedComponentsReplacer,
-    serializedComponentsReviver,
-} from "./src/Core/utils/serializedStateProcessing";
-export { returnAllPossibleVariants } from "./src/Core/utils/returnAllPossibleVariants";
-export { CodeMirror } from "@doenet/codemirror";
