+++ packages/doenetml/src/index.js
@@ -1,0 +1,17 @@
+export { DoenetML } from "./DoenetML";
+
+export { mathjaxConfig } from "./Core/utils/math";
+export { cidFromText } from "./Core/utils/cid";
+export { retrieveTextFileForCid } from "./Core/utils/retrieveTextFile";
+export {
+    calculateOrderAndVariants,
+    determineNumberOfActivityVariants,
+    parseActivityDefinition,
+    returnNumberOfActivityVariantsForCid,
+} from "./utils/activityUtils";
+export {
+    serializedComponentsReplacer,
+    serializedComponentsReviver,
+} from "./Core/utils/serializedStateProcessing";
+export { returnAllPossibleVariants } from "./Core/utils/returnAllPossibleVariants";
+export { CodeMirror } from "@doenet/codemirror";
