+++ packages/doenetml/tsconfig.json
@@ -4,7 +4,7 @@
         "outDir": "./dist/",
         "rootDir": "./src"
     },
-    "include": ["./**/*.ts", "./**/*.tsx"],
+    "include": ["./**/*.ts", "./**/*.tsx", "./**/*.js", "./**/*.jsx"],
     "exclude": ["vite.config.ts"],
     "references": []
 }
