+++ packages/test-viewer/index-inline-worker.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="shortcut icon" href="/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>DoenetML Viewer</title>
+    <script src="https://www.youtube.com/iframe_api"></script>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main-inline-worker.jsx"></script>
+  </body>
+</html>
