+++ packages/test-viewer/src/main-inline-worker.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import TestViewer from "./test/testViewer";
-import { DoenetML } from "@doenet/doenetml";
+import { DoenetML } from "@doenet/doenetml/doenetml-inline-worker.js";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
     <TestViewer DoenetML={DoenetML} />,
